refactor: reorganize processing module structure and update imports
