+++ tests/test_crud/test_importation.py
@@ -1,5 +1,5 @@
 from app.crud.importation import create_importation, get_by_year_and_country_and_path
-from app.models.processing import ImportationCreate
+from app.processing.models import ImportationCreate
 
 
 def test_create_and_get(db_session):
